lib: Move types & methods to own files.
