+++ lib/diff_rez_message.go
@@ -1,0 +1,27 @@
+package sous
+
+import "github.com/opentable/sous/util/logging"
+
+type diffRezMessage struct {
+	resolution *DiffResolution
+	callerInfo logging.CallerInfo
+}
+
+func (msg diffRezMessage) DefaultLevel() logging.Level {
+	return logging.WarningLevel
+}
+
+func (msg diffRezMessage) Message() string {
+	return string(msg.resolution.Desc)
+}
+
+func (msg diffRezMessage) EachField(f logging.FieldReportFn) {
+	f("@loglov3-otl", "sous-diff-resolution")
+	msg.callerInfo.EachField(f)
+	f("sous-deployment-id", msg.resolution.DeploymentID.String())
+	f("sous-manifest-id", msg.resolution.ManifestID.String())
+	f("sous-resolution-description", string(msg.resolution.Desc))
+	marshallable := buildMarshableError(msg.resolution.Error.error)
+	f("sous-resolution-errortype", marshallable.Type)
+	f("sous-resolution-errormessage", marshallable.String)
+}
